Checkout latest tag of dependencies
For the tests, instead of only using the latest tag for
dbus-client-gen and HEAD for dbus-signature-pyparsing,
dbus-python-client-gen, and into-dbus-python, use a "for"
loop that checks out the latest tag for all four.

While here, also relocate the comment for setting the PYTHONPATH,
moving it right above the relevant block.  Also, add newlines
for clarity.

Signed-off-by: Bryan Gurney <bgurney@redhat.com>