ci: add non-rust variant of stratisd test
Add a variant of the stratisd test that performs only the non-rust
tests.  This test script does not require any real test devices,
and can be executed in the same manner as the stratisd.sh test
script.

The stratisd.sh test script will not be modified now, but the
non-rust tests can be removed later.

Signed-off-by: Bryan Gurney <bgurney@redhat.com>