ci: add non-rust variant of stratisd test
Add a variant of the stratisd test that performs only the non-rust
tests.  This test script does not require any real test devices,
and can be executed in the same manner as the stratisd.sh test
script.

The stratisd.sh test script will not be modified now, but the
non-rust tests can be removed later.

Additional modifications from the original stratisd.sh:

- Remove the unused TEST_BLOCKDEVS_FILE variable.
- Fix a typo in the "Required $STRATISD not found" echo.
- Remove a superfluous cd from the old "git pull" of dependencies.

Signed-off-by: Bryan Gurney <bgurney@redhat.com>