ci: add teardown script for stratisd test
If the "stratisd.sh test-real" test is terminated in mid-run,
there may be a remnant Stratis pool left on the test devices,
potentially with a thin-pool that was left suspended.

This test in particular seems to leave a device behind in that state:

engine::strat_engine::pool::tests::real_test_add_datadevs

The "thinpool-pool" device undergoes a high frequency of
suspend-reload-resume activity during this test.  If it isn't
resumed, it may cause a process to go into uninterruptible sleep,
resulting in devices that can't be removed, and requiring a restart
of the test system.

The script terminates any test processes (including "stratisd --sim"
processes left over from devicemapper-rs tests), resumes the testing
"thinpool-pool" devices, and removes the devices in the Stratis
pool, top to bottom.  After that, any block devices specified in the
test_config.json file are wiped clean of signatures, via "wipefs -a".

Signed-off-by: Bryan Gurney <bgurney@redhat.com>