ci blackbox: echo recent commits for stratisd/stratis-cli repo
When the blackbox test clones the stratisd and stratis-cli
repositories, there isn't any information in stdout regarding the
latest commit used (in the master branch, which is built by default
during the blackbox test).

Add a "git log" command that prints the short commit hash, author
date in ISO-8601 format with local time zone, and subject.
(This is similar to "git log --oneline", but with date/time info,
to provide temporal context.)

The 10 most recent commits are printed in reverse chronological
order.

Signed-off-by: Bryan Gurney <bgurney@redhat.com>