ci blackbox: print committer date in most recent commits
In the stratis-blackbox-run.sh script, there are commands that print
the header "Most recent commits for (repo)", followed by a "git log"
command executed in the repository directory.  This helps to log the
most recent commits in the stratisd and stratis-cli repositories used
to build the blackbox packages, which may help a tester to determine
if two different blackbox runs were executed on the same code, or
before or after a certain amount of commits.

Instead of using the author date in the custom "git log --format"
command, use the committer date; this will provide the date of
the commit, which will descend in chronological order.

Also, print the most recent 20 commits; due to the pull-request
progression of the commit history, this will (usually) allow the
last 10 commits to be seen, instead of the last 5 commits.

Signed-off-by: Bryan Gurney <bgurney@redhat.com>