Add coverage requirement to cli Jenkins tests
Use the no html coverage target in the stratis-cli Makefile. It will very
seldom be of use to generate the html, since most of the time the tests
will pass, and since the report also indicates where code coverage is
missing. It might someday be sort of annoying to have to clean up all the
html, so let's not generate it until we know we need to.

Require python3-coverage in setup script.

Signed-off-by: mulhern <amulhern@redhat.com>