ci teardown: unmount stale test dirs after resume
If there's a test that doesn't relinquish control of a
"stratis_testing" test directory in /tmp, it may occur while the
thin pool under test is suspended.  In this case, the test may
hang, and the subsequent attempts to tear down the device-mapper
devices may fail due to the open mount.

After resuming all thinpool-pool devices, print the name of
any "/tmp/stratis_testing..." directories, and unmount them.

Signed-off-by: Bryan Gurney <bgurney@redhat.com>