Reverse order in which blackbox tests are run
I prefer to see the stratisd tests run first, as they test only stratisd,
while the stratis-cli tests test both the CLI and the daemon.

If there's a real failure in a test, it may leave state that causes spurious
failures of subsequent tests. I'ld like to have the stratisd tests have the
chance to have the first failure, because that failure is likely to be more
real than all subsequent failures, and a failure in the daemon is really
more fundamental than one in the CLI.

Signed-off-by: mulhern <amulhern@redhat.com>