Allow for partial versions in Cargo.toml
In stratisd, we always use complete versions for specifying dependenceis.
But sometimes we want to run this script on other upstream projects, which
do not, or did not always, use complete versions.

Add a command-line option to relax the version format restriction.

The partial flag is deprecated and will be removed in semantic_version
3.0. At that point, we'll have to do something like switch to using the
packaging package or construct complete versions from partial versions
before passing to the Version constructor or some other solution.

Signed-off-by: mulhern <amulhern@redhat.com>