Use ExclusiveArch only if RHEL
Fedora Rust SIG believes that ExclusiveArch is no longer needed for
Rust packages in Fedora and no longer includes ExclusiveArch
specification in automatically generated Rust spec files.

The reasoning is that Rust is now supported in all architectures that
Fedora supports, because rpm now has Rust components and rpm must be
available on all architectures that Fedora supports.

This convention was introduced in rust2rpm v23, which was released
October 2022.

Signed-off-by: mulhern <amulhern@redhat.com>