generate_sourcerpms: Add CentOS Stream specific spec files
The stratis-cli upstream project now uses a new setup.cfg file that
is too new for CentOS Stream to be able to properly determine the
packages necessary for building.  Therefore, add a new copy of the
stratis-cli.spec file, as stratis-cli-centosstream.spec, which will
be copied in when using the "centos-stream" script target.

Signed-off-by: Bryan Gurney <bgurney@redhat.com>