Revert stratisd.spec file 3.5.8 patch configuration
While the updates to the stratisd.spec file for patching the vendor
tar file were successful in CentOS Stream for stratisd-3.5.8, trying
to build on the latest stratisd version with the vendor file patch
approach fails, while the older method succeeds.

Revert the change, but keep it in the commit history in case it's
needed later.

Signed-off-by: Bryan Gurney <bgurney@redhat.com>