Drop the license summary
It is just visual clutter in the build log output. The cargo_license
macro, which is invoked just below the removed line actually generates a
list of dependency licenses and places it in a file in the rpm. We can make
that rpm in a scratch build and then extract it and post-process its
info however we like.

Signed-off-by: mulhern <amulhern@redhat.com>