Remove update_workflow_config.py
We should focus on using sd to replace values in the workflow
configurations.

The value added by the scripts is now just knowledge of the keys. That
information is easily extracted from the workflow files themselves, if
we forget.

This way, we do not have to extend the script to also update Python
versions in the configuration, we will just add similarly formatted keys
and use sd for that, also.

Signed-off-by: mulhern <amulhern@redhat.com>